+++ react-components/src/index.ts
@@ -55,3 +55,7 @@ export * from './components/Theme';
 
 export * from './components/Gantt';
 export * from './components/Gantt/model/public-api';
+
+export * from './components/Modal';
+
+export * from './components/Badge';